chore(suite-native): Mobile Sell: refactor useTradingAnalyticReportCallback types
