+++ suite-native/module-trading/src/components/fees/__tests__/FeePickerCard.comp.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TradingType } from '@suite-common/trading';
+import { TradingExchangeType, TradingSellType } from '@suite-common/trading';
 import { NetworkSymbol } from '@suite-common/wallet-config';
 import { AccountKey } from '@suite-common/wallet-types';
 import { renderWithStoreProviderAsync, userEvent } from '@suite-native/test-utils';
@@ -30,7 +30,7 @@ describe('FeePickerCard', () => {
         trade: undefined,
         accountKey: 'btc1' as AccountKey,
         symbol: 'btc' as NetworkSymbol,
-        tradingType: 'buy' as TradingType,
+        tradingType: 'exchange' as TradingExchangeType | TradingSellType,
     };
 
     const createPreloadedState = (fee: string) => ({
@@ -104,7 +104,7 @@ describe('FeePickerCard', () => {
 
         expect(mockNavigate).toHaveBeenCalledWith('TradingFees', {
             accountKey: 'btc1',
-            tradingType: 'buy',
+            tradingType: 'exchange',
         });
     });
